+++ app/src/main/java/com/example/catedu/MainActivity.java
@@ -21,7 +21,6 @@ public class MainActivity extends AppCompatActivity {
 
     /**
      * MainActivity 创建时的操作
-     * @param Bundle savedInstanceState
      */
     @Override
     protected void onCreate(Bundle savedInstanceState) {