+++ Change Log.md
@@ -14,3 +14,4 @@
 ## 8.12
 
 - Add annotations.
+- Realize http connection and post request (login) .